Recommendations Endpoint Completed
